initial implementation of journaling system functionality
